Veyoo/backend/imagen-service/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-generativeai
google-cloud-storage
python-dotenv